Move TestESRInterconnectGuard to test_cluster_app_algorithms.py using session-scoped fixture
Agent-Logs-Url: https://github.com/gschivley/PowerGenome-tools/sessions/1f85bb61-89ec-4bf0-a42b-1305bcd39609
