+++ tests/test_cluster_app.py
@@ -11,12 +11,8 @@ NOTE: Real-import tests (color utils, graph algorithms, clustering, ESR policy,
 plant clustering, etc.) live in test_cluster_app_algorithms.py.
 """
 
-import importlib.util
 import math
 import re
-import sys
-from pathlib import Path
-from unittest.mock import MagicMock
 
 import numpy as np
 import pandas as pd
@@ -2226,218 +2222,6 @@ SingleRegion:
         assert "10 regions with 20 BAs" in msg
         assert len(state.manual_regions) == 10
 
-
-# ============================================================================
-# ESR Interconnect Guard Tests
-# ============================================================================
-
-
-@pytest.fixture()
-def _cluster_app_for_esr():
-    """Load cluster_app with mocked js/PyScript deps (function-scoped)."""
-    module_names = [
-        "js",
-        "pyodide",
-        "pyodide.ffi",
-        "renewables_utils",
-        "fast_interconnection",
-        "fast_interconnection.fast_assign",
-        "fast_interconnection.resource_groups",
-        "cluster_app",
-    ]
-    original_modules = {name: sys.modules.get(name) for name in module_names}
-    web_dir = None
-
-    try:
-        mock_js = MagicMock()
-        mock_ffi = MagicMock()
-        mock_ffi.create_proxy = lambda x: x
-        mock_ffi.to_js = lambda x: x
-        mock_ffi.JsProxy = object
-
-        sys.modules["js"] = mock_js
-        sys.modules["pyodide"] = MagicMock()
-        sys.modules["pyodide.ffi"] = mock_ffi
-        sys.modules["renewables_utils"] = MagicMock()
-        sys.modules["fast_interconnection"] = MagicMock()
-        sys.modules["fast_interconnection.fast_assign"] = MagicMock()
-        mock_rg = MagicMock()
-        mock_rg.DEFAULT_PROFILE_PATHS = {}
-        mock_rg.build_assigned_df = MagicMock(return_value=None)
-        mock_rg.build_resource_group_json = MagicMock(return_value={})
-        sys.modules["fast_interconnection.resource_groups"] = mock_rg
-
-        mock_js.L = MagicMock()
-        mock_js.document = MagicMock()
-        mock_js.window = MagicMock()
-        mock_js.fetch = MagicMock()
-        mock_js.Uint8Array = MagicMock()
-        mock_js.globalThis = MagicMock()
-
-        web_dir = Path(__file__).parent.parent / "web"
-        sys.path.insert(0, str(web_dir))
-        module_path = web_dir / "cluster_app.py"
-        spec = importlib.util.spec_from_file_location("cluster_app", module_path)
-        module = importlib.util.module_from_spec(spec)
-        sys.modules["cluster_app"] = module
-        assert spec.loader is not None
-        spec.loader.exec_module(module)
-
-        yield module
-    finally:
-        if web_dir is not None and str(web_dir) in sys.path:
-            sys.path.remove(str(web_dir))
-        for name, original in original_modules.items():
-            if original is None:
-                sys.modules.pop(name, None)
-            else:
-                sys.modules[name] = original
-
-
-class TestESRInterconnectGuard:
-    """Regression tests for the cross-interconnect guard in generate_resource_tags_settings().
-
-    Before the fix, WY1 (SD/western) was incorrectly tagged for a WI ESR constraint
-    because SD appears in the rectable for WI, ignoring that p32's SD BA is in the
-    western interconnect while WI's policy state is eastern.
-    """
-
-    @pytest.fixture()
-    def app(self, _cluster_app_for_esr):
-        """Inject minimal AppState for cross-interconnect ESR tagging scenarios."""
-        m = _cluster_app_for_esr
-
-        # p32: SD BA in the western interconnect.
-        # p46: WI BA in the eastern interconnect.
-        m.state.hierarchy_df = pd.DataFrame(
-            {
-                "ba": ["p32", "p46"],
-                "st": ["SD", "WI"],
-                "interconnect": ["Western", "Eastern"],
-            }
-        )
-
-        # rectable: WI policy row only.
-        # rectable.loc["WI", "SD"] = 2 → SD can satisfy WI's RPS per rectable.
-        # WI→WI handled by the same-state shortcut inside can_generator_satisfy_policy.
-        m.state.rectable_df = pd.DataFrame(
-            {"WI": [1], "SD": [2]},
-            index=["WI"],
-        )
-
-        # WY1 holds only the SD/western BA; eastern2 holds only the WI/eastern BA.
-        m.state.region_aggregations = {
-            "WY1": ["p32"],
-            "eastern2": ["p46"],
-        }
-
-        # One RPS constraint whose policy state is WI.
-        # esr_policy_states values are lowercase (matching get_states_in_region output).
-        m.state.esr_map = {"ESR_1": {}}
-        m.state.esr_type_map = {"ESR_1": "RPS"}
-        m.state.esr_policy_states = {"ESR_1": {"wi"}}
-        m.state.esr_rps_techs = {"Wind"}
-        m.state.esr_ces_techs = set()
-
-        # No new resources — keeps output minimal and CCS branches silent.
-        m.state.modified_new_resources = {}
-        m.state.new_resources = []
-
-        return m
-
-    def test_western_sd_region_excluded_from_eastern_wi_esr(self, app):
-        """WY1 (SD BA, western interconnect) must NOT be tagged for the WI ESR.
-
-        SD can satisfy WI's RPS per the rectable (value=2), but p32's SD BA
-        sits in the western interconnect while WI's policy state is eastern.
-        The interconnect guard must prevent the tag from being assigned.
-        """
-        result = yaml.safe_load(app.generate_resource_tags_settings())
-        regional = result.get("regional_tag_values", {})
-
-        wy1_esr_tags = regional.get("WY1", {})
-        assert "ESR_1" not in wy1_esr_tags, (
-            "WY1 (SD, western) must not receive the WI ESR_1 tag "
-            "even though SD appears in the rectable for WI"
-        )
-
-    def test_eastern_wi_region_included_in_wi_esr(self, app):
-        """eastern2 (WI BA, eastern interconnect) MUST be tagged for the WI ESR.
-
-        A WI generator shares the eastern interconnect with the WI policy state
-        and always satisfies its own state's ESR constraint.
-        """
-        result = yaml.safe_load(app.generate_resource_tags_settings())
-        regional = result.get("regional_tag_values", {})
-
-        assert (
-            "eastern2" in regional
-        ), "eastern2 (WI, eastern) must appear in regional_tag_values"
-        assert "ESR_1" in regional["eastern2"], "eastern2 must carry the WI ESR_1 tag"
-
-    def test_split_state_minority_interconnect_not_excluded(self, _cluster_app_for_esr):
-        """A region whose BA is in the *minority* interconnect for its state must not
-        be incorrectly excluded from ESR tagging when it shares an interconnect with
-        the policy state.
-
-        Scenario:
-          - SD has 3 eastern BAs (p_sd_e1, p_sd_e2, p_sd_e3) and 1 western BA (p_sd_w).
-            Under a majority-based state→interconnect mapping, SD maps to "Eastern".
-          - Region "western_sd" contains only p_sd_w (the minority/western SD BA).
-          - Policy state "nd" (ND) is in the western interconnect.
-          - rectable.loc["ND", "SD"] = 1 → SD generators can satisfy ND's RPS.
-          - The interconnect guard must NOT exclude "western_sd" because p_sd_w's actual
-            interconnect (Western) matches ND's interconnect (Western).
-          - A majority-based guard would incorrectly see SD as "Eastern" and block the tag.
-        """
-        m = _cluster_app_for_esr
-
-        # SD has 3 eastern BAs (majority eastern) + 1 western BA (minority western).
-        # ND has 1 western BA.
-        m.state.hierarchy_df = pd.DataFrame(
-            {
-                "ba": ["p_sd_e1", "p_sd_e2", "p_sd_e3", "p_sd_w", "p_nd_w"],
-                "st": ["SD", "SD", "SD", "SD", "ND"],
-                "interconnect": ["Eastern", "Eastern", "Eastern", "Western", "Western"],
-            }
-        )
-
-        # rectable: ND policy row only.
-        # rectable.loc["ND", "SD"] = 1 → SD generators can satisfy ND's RPS.
-        # ND→ND handled by the same-state shortcut.
-        m.state.rectable_df = pd.DataFrame(
-            {"ND": [1], "SD": [1]},
-            index=["ND"],
-        )
-
-        # "western_sd" holds only the minority-interconnect SD BA.
-        # "nd_region" holds the ND/western BA.
-        m.state.region_aggregations = {
-            "western_sd": ["p_sd_w"],
-            "nd_region": ["p_nd_w"],
-        }
-
-        # One RPS constraint whose policy state is ND (western interconnect).
-        m.state.esr_map = {"ESR_1": {}}
-        m.state.esr_type_map = {"ESR_1": "RPS"}
-        m.state.esr_policy_states = {"ESR_1": {"nd"}}
-        m.state.esr_rps_techs = {"Wind"}
-        m.state.esr_ces_techs = set()
-
-        m.state.modified_new_resources = {}
-        m.state.new_resources = []
-
-        result = yaml.safe_load(m.generate_resource_tags_settings())
-        regional = result.get("regional_tag_values", {})
-
-        assert "western_sd" in regional, (
-            "western_sd (SD minority-western BA) must appear in regional_tag_values "
-            "because its actual BA interconnect (Western) matches ND's interconnect"
-        )
-        assert "ESR_1" in regional["western_sd"], (
-            "western_sd must carry the ND ESR_1 tag — the BA-level interconnect guard "
-            "must not exclude it based on SD's majority (Eastern) interconnect"
-        )
 
 
 if __name__ == "__main__":
